Factor out Player Management from Game Manager
